feat: optimize app deleted
